refactor: remove circular dependencies
