--- src/components/AccordionButton/AccordionButton.module.css
@@ -1,82 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # ACCORDION BUTTON
-\*------------------------------------*/
-
-/**
- * Accordion Button, wraps the heading and open indicator icon.
- */
-.accordion-button {
-  display: flex;
-  justify-content: space-between;
-
-  border: 0;
-  border-radius: 0;
-
-  padding: var(--eds-size-2) var(--eds-size-1);
-  height: 3.375rem;
-}
-
-/**
- * Small variant.
- */
-.accordion-button--sm {
-  padding: var(--eds-size-1);
-  height: 2.25rem;
-}
-
-/**
- * Outline variant.
- */
-.accordion-button--outline {
-  padding: var(--eds-size-2) var(--eds-size-3);
-}
-
-/**
- * Small outline variant.
- */
-.accordion-button--sm.accordion-button--outline {
-  padding: var(--eds-size-1) var(--eds-size-2);
-}
-
-/**
- * The heading text.
- */
-.accordion-button__heading {
-  @mixin eds-theme-typography-body-text-sm;
-  color: var(--eds-theme-color-text-neutral-strong);
-}
-
-/**
- * Small variant of the heading text.
- */
-.accordion-button__heading--sm {
-  @mixin eds-theme-typography-body-text-xs;
-}
-
-/**
- * Expand more (chevron) icon indicates open or closed status.
- *
- * This nonrotated icon points down and represents closed status.
- */
-.accordion-button__icon {
-  transform: rotate(0);
-}
-/**
- * This rotated icon points up and represents open status.
- */
-.accordion-button__icon--open {
-  transform: rotate(180deg);
-}
-
-/**
- * Animates the icon rotation when opening and closing.
- */
-.accordion-button > .accordion-button__icon {
-  transition: transform var(--eds-anim-move-medium) var(--eds-anim-ease);
-
-  @media screen and (prefers-reduced-motion) {
-    transition: none;
-  }
-}
