--- src/components/AccordionButton/index.ts
@@ -1,1 +1,0 @@
-export { AccordionButton as default } from './AccordionButton';
