--- src/components/AccordionPanel/AccordionPanel.module.css
@@ -1,41 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # ACCORDION PANEL
-\*------------------------------------*/
-
-/**
- * Accordion Panel that expands and collapses.
- * Reveals itself below the associated Accordion Button.
- */
-.accordion-panel {
-  padding-left: var(--eds-size-1);
-  padding-right: var(--eds-size-3);
-  padding-bottom: 1.625rem;
-
-  @mixin eds-theme-typography-body-text-sm;
-  color: var(--eds-theme-color-text-neutral-strong);
-}
-
-/**
- * Small variant.
- */
-.accordion-panel--sm {
-  @mixin eds-theme-typography-body-text-xs;
-}
-
-/**
- * Outline variant.
- */
-.accordion-panel--outline {
-  padding-left: var(--eds-size-3);
-  padding-right: var(--eds-size-5);
-}
-
-/**
- * Small outline variant.
- */
-.accordion-panel--sm.accordion-panel--outline {
-  padding-left: var(--eds-size-2);
-  padding-right: var(--eds-size-4);
-}
