--- src/components/AccordionPanel/index.ts
@@ -1,1 +1,0 @@
-export { AccordionPanel as default } from './AccordionPanel';
