--- src/components/AccordionRow/AccordionRow.module.css
@@ -1,19 +1,0 @@
-/*------------------------------------*\
-    # ACCORDION ROW
-\*------------------------------------*/
-
-/**
- * Accordion Row houses one Accordion Button subcomponent and its relevant Accordion Panel subcomponent.
- */
-.accordion-row {
-  border-bottom: var(--eds-border-width-sm) solid
-    var(--eds-theme-color-border-neutral-subtle);
-}
-
-/**
- * Variant with outline.
- * Removes bottom border on the last Accordion Row so there aren't two borders at the bottom of the Accordion.
- */
-.accordion-row--outline:last-of-type {
-  border-bottom: 0;
-}
