--- src/components/AccordionRow/index.ts
@@ -1,1 +1,0 @@
-export { AccordionRow as default } from './AccordionRow';
